fixed carousel, skeleton and timer
